+++ .gitignore
@@ -1,1 +1,3 @@
 venv
+tests/__pycache__/*
+pyclub/__pycache__/*
